# backend/service.core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Order:
    properties:
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/entity.OrderItem'
        type: array
      restaurant_id:
        type: string
      status:
        type: string
      table_num:
        type: integer
      total_cost:
        type: number
      user_id:
        type: string
    type: object
  entity.OrderItem:
    properties:
      item_id:
        type: string
      quantity:
        type: integer
    type: object
  entity.OrderItemResult:
    properties:
      item_id:
        type: string
      item_name:
        type: string
      item_price:
        type: integer
      quantity:
        type: integer
    type: object
  entity.OrderResult:
    properties:
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/entity.OrderItemResult'
        type: array
      restaurant_id:
        type: string
      status:
        type: string
      table_num:
        type: integer
      total_cost:
        type: number
      user_id:
        type: string
    type: object
...
  errors.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  orders.CreateOrderItemRequest:
    properties:
      item_id:
        type: string
      order_id:
        type: string
      quantity:
        type: integer
      restaurant_id:
        type: string
    type: object
  orders.CreateOrderRequest:
    properties:
      items:
        additionalProperties:
          type: integer
        type: object
      restaurant_id:
        type: string
      table_num:
        type: integer
    type: object
  orders.GetItemResponse:
    properties:
      item:
        $ref: '#/definitions/entity.OrderResult'
        type: object
    type: object
  orders.GetListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.OrderResult'
        type: array
    type: object
  orders.ItemResponse:
    properties:
      item:
        $ref: '#/definitions/entity.Order'
        type: object
    type: object
  orders.MoveOrderRequest:
    properties:
      new_table_number:
        type: integer
      restaurant_id:
        type: string
    type: object
  orders.OrderItemResponse:
    properties:
      item:
        $ref: '#/definitions/entity.OrderItem'
        type: object
    type: object
  orders.UpdateOrderItemRequest:
    $ref: '#/definitions/orders.CreateOrderItemRequest'
  orders.UpdateOrderRequest:
    properties:
      items:
        additionalProperties:
          type: integer
        type: object
      restaurant_id:
        type: string
      status:
        type: string
      table_num:
        type: integer
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Restaurants and customers and other great things!
  license: {}
  title: COMP3900 - JAMAR
  version: "0.1"
paths:
...
  /order:
    get:
      parameters:
      - description: If supplied by user, only return orders made by this table
        in: query
        name: res_id
        required: true
        type: string
      - description: If supplied by user, only return orders made by this specific user
        in: query
        name: table_num
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.GetItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Gets a list of orders depending on the query parameters
      tags:
      - Orders
  /orders:
    get:
      parameters:
      - description: If supplied by user, only return orders made by this restaurant
        in: query
        name: res_id
        type: string
      - description: If supplied by user, only return orders made by this specific user
        in: query
        name: usr_id
        type: string
      - description: If supplied, return a list of order based on the status code. Should only be used by restaurant
        in: query
        name: status
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.GetListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Gets a list of orders depending on the query parameters
      tags:
      - Orders
    post:
      parameters:
      - description: The new order's details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.CreateOrderRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Creates a order. Used by both customer and users
      tags:
      - Orders
  /orders/{ord_id}:
    post:
      parameters:
      - description: The id of the order
        in: path
        name: ord_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.OrderItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Creates a new order item in an order
      tags:
      - Orders
    put:
      parameters:
      - description: The id of the order
        in: path
        name: ord_id
        required: true
        type: string
      - description: The order item's details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.UpdateOrderItemRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.OrderItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Updates an order item in an order.
      tags:
      - Orders
  /orders/{ord_id}/cancel:
    post:
      parameters:
      - description: The id of the order
        in: path
        name: ord_id
        required: true
        type: string
      - description: The order's updated details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.UpdateOrderRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Sets an order status to "canceled"
      tags:
      - Orders
  /orders/{ord_id}/complete:
    post:
      parameters:
      - description: The id of the order
        in: path
        name: ord_id
        required: true
        type: string
      - description: The order's updated details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.UpdateOrderRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Sets an order status to "completed"
      tags:
      - Orders
  /orders/{ord_id}/cook:
    post:
      parameters:
      - description: The id of the order
        in: path
        name: ord_id
        required: true
        type: string
      - description: The order's updated details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.UpdateOrderRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Sets an order status to "served"
      tags:
      - Orders
  /orders/{ord_id}/items/{itm_id}:
    delete:
      parameters:
      - description: The id of the order
        in: path
        name: ord_id
        required: true
        type: string
      - description: The id of the item in the order
        in: path
        name: itm_id
        required: true
        type: string
      - description: The order item's details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.UpdateOrderItemRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Deletes an order item from an order
      tags:
      - Orders
  /orders/{ord_id}/move:
    put:
      parameters:
      - description: Move an order to different table
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.MoveOrderRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Updates an order
      tags:
      - Orders
...
swagger: "2.0"